link container refactoring, improved retry behaviour for failed port messages
