+++ src/crazyradio/deps/libcrtp/CMakeLists.txt
@@ -5,5 +5,6 @@ project(libcrtp)
 include_directories(include)
 add_library(libcrtp 
         CrtpPacketQueue.cpp
-        CrtpLink.cpp)
+        CrtpLink.cpp
+        CrtpLinkContainer.cpp)
