+++ src/crazyradio/deps/libcrtp/include/libcrtp/CrtpLink.hpp
@@ -1,14 +1,12 @@
 #pragma once
-
 #include <stdint.h>
-
 #include <map>
-#include <mutex>
 
 #include "libcrtp/CrtpPacketQueue.hpp"
 #include "libcrtp/CrtpPacket.hpp"
 
 namespace libcrtp {
+
 struct CrtpLinkIdentifier
 {
     uint8_t channel;
@@ -63,10 +61,15 @@ class CrtpLink
         */
         void notifySuccessfullPortMessage(CrtpPort port);
 
+        /** 
+         * Notifies about a failed nullpacket attempt, returns true if link shall die.
+         */
+        bool notifyFailedNullpacket();
+
         /**
          * Notifies about a failed send attempt, returns true if link shall die
         */
-        bool notifyFailedMessage();
+        bool notifyFailedPortMessage();
 
         /**
          * This is called before decontrstuction.
@@ -96,91 +99,32 @@ class CrtpLink
 
     private: 
         void onSuccessfullMessage();
+        bool onFailedMessage();
         
-    // Configurable parameters
-    private: 
-        uint8_t m_failedMessagesMaximum;
-        uint32_t m_relaxationPeriodMs;
-        uint32_t m_lastSuccessfullMessageTimeout;
-        uint32_t m_failedMessageRetryTimeout;
-
-    private: 
-        uint8_t m_failedMessagesCount;
-        uint32_t m_relaxationCountMs;
-        uint32_t m_lastSuccessfullMessageTime;
-        
+    private:
         std::map<CrtpPort, CrtpPacketQueue> m_crtpPortQueues;
-
-        uint64_t m_linkQuality; // 64 bits of failed and successful messages (bits)
 
         uint8_t m_channel;
         uint64_t m_address;
         uint8_t m_datarate; 
         bool m_isBroadcast;
-};
-
-
-class CrtpLinkContainer
-{
-    public:
-        CrtpLinkContainer();
-        virtual ~CrtpLinkContainer();
-
-        /**
-         * Adds Link to Container, if already present will not overwrite
-         */
-        void addLink(uint8_t channel, uint64_t address, uint8_t datarate);
-
-        /**
-         * Removes a link from the container.
-         * Returns true if the link was removed, false if it was not found.
-         * The callbacks are returned to the caller, they should probably be called with a nullpacket.
-         */
-        bool removeLink(CrtpLinkIdentifier * link, std::vector<CrtpResponseCallback>& callbacks);
-
-        /**
-         * Returns true if the link was found and copied to the link identifier.
-         * If the link is not found, false is returned.
-         */
-        bool getLinkIdentifier(CrtpLinkIdentifier * link, uint8_t channel, uint64_t address) const;
-
-        /**
-         * Returns true if a link with the highest priority port is found.
-         * The link identifier is copied to the link parameter and the port is set to the highest priority port of the link.
-         */
-        bool getHighestPriorityLink(CrtpLinkIdentifier * link, CrtpPort * port) const;
-
-        /**
-         * Returns a random relaxed link which is not a broadcast link.
-         * If no such link is found, false is returned.
-         */
-        bool getRandomRelaxedNonBroadcastLink(CrtpLinkIdentifier * link) const;
-
-        /**
-         * Get connection statistics for all links.
-         */
-        void getConnectionStats(std::vector<CrtpLinkIdentifier>& links, std::vector<double>& quality) const;
-
-        /**
-        * In order to call link functions with link identifiers we need these wrappers
-        */
-        void tickLinksMs(uint8_t ms);
-        
-        void linkAddPacket(CrtpLinkIdentifier * link, CrtpPacket * packet, CrtpResponseCallback callback);
-        bool linkGetHighestPriorityPacket(CrtpLinkIdentifier * link, CrtpPacket * packet);
-        bool linkReleasePacket(CrtpLinkIdentifier  * link_id, CrtpPacket * responsePacket, CrtpResponseCallback & callback);
-
-        void linkNotifySuccessfullNullpacket(CrtpLinkIdentifier * link_id);
-        void linkNotifySuccessfullPortMessage(CrtpLinkIdentifier * link_id, CrtpPort port);
-        bool linkNotifyFailedMessage(CrtpLinkIdentifier * link_id);
-
+    
+    // Configurable parameters
     private: 
-        void copyLinkIdentifier(CrtpLinkIdentifier * from_link, CrtpLinkIdentifier * to_link) const;
-        void linkToIdentifier(const CrtpLink * link, CrtpLinkIdentifier * link_id) const;
-        bool linkFromIdentifier(CrtpLink ** link, CrtpLinkIdentifier * link_id);
-        std::map<std::pair<uint8_t, uint64_t>, libcrtp::CrtpLink> m_links;
+        uint8_t m_failedMessagesMaximum;
+        uint32_t m_nullpacketPeriodMs;
+        uint32_t m_lastSuccessfullMessageTimeoutMs;
+        uint32_t m_failedMessageRetryTimeoutMs;
+    
+    // Internal state
+    private: 
+        uint8_t m_failedMessagesCount;
+        uint32_t m_timeSinceLastSuccessfullMessageMs;
 
-        mutable std::mutex m_linksMutex;
+        uint8_t m_failedPortMessagesCount;
+        uint32_t m_timeSinceLastNullpacketMs;
+        uint32_t m_timeSinceLastFailedPortMessageMs;
+        uint64_t m_linkQuality; // 64 bits of failed and successful messages (bits)
 };
 
 } // namespace libcrtp
